test_requirements.txt
tox
coverage
# django-coverage
flake8
# factory_boy
# mock
selenium

django-cms==3.3.2
django-ckeditor
django-filer